src/data/model/rag_model.py
from typing import Dict, Optional

import nvidia.nim as nim
from transformers import AutoModel, AutoTokenizer

class RAGModel:
    def __init__(self, config: Dict):
        self.config = config
        self.retriever = self._initialize_retriever()
        self.generator = self._initialize_generator()
        
    def _initialize_retriever(self):
        """
        Initialize the retriever component using NVIDIA NIM
        """
        retriever_config = self.config['retriever']
        # Implementation using NVIDIA NIM
        return None
        
    def _initialize_generator(self):
        """
        Initialize the generator component using NVIDIA NIM
        """
        generator_config = self.config['generator']
        # Implementation using NVIDIA NIM
        return None
        
    def generate(self, query: str, num_returns: int = 1):
        """
        Generate response using RAG
        """
        # Implementation for generation
        pass 