requirements.txt
nvidia-nim
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0